[Rename] 파일 이름 변경
- userGetDetailCommon -> useGetDetailCommon
- common뿐만 아니라 intro 정보도 가져옴

# 167
